+++ client/src/utils/cleanForSlug.test.js
@@ -1,12 +1,6 @@
-window.$ = require('../../../../wagtail/admin/static_src/wagtailadmin/js/vendor/jquery-3.5.1.min');
 // eslint-disable-next-line no-unused-expressions
-require('../../../../wagtail/admin/static_src/wagtailadmin/js/vendor/urlify').default;
-
-window.comments = {
-  getContentPath: jest.fn(),
-};
-
-const cleanForSlug = require('./page-editor').cleanForSlug;
+require('../../../wagtail/admin/static_src/wagtailadmin/js/vendor/urlify').default;
+import { cleanForSlug } from './cleanForSlug';
 
 describe('page-editor tests', () => {
   describe('cleanForSlug without unicode slugs enabled', () => {
